allows sortable for nested columns (should be use carefully); extracts Computor from Data; improves excel export -> fetcher needs to be redesiged
