+++ src/Services/Template/Validators/Columns/Meta.php
@@ -2,7 +2,6 @@
 
 namespace LaravelEnso\Tables\Services\Template\Validators\Columns;
 
-use Illuminate\Support\Str;
 use LaravelEnso\Helpers\Services\Obj;
 use LaravelEnso\Tables\Attributes\Column as Attributes;
 use LaravelEnso\Tables\Exceptions\Meta as Exception;
@@ -17,13 +16,6 @@ class Meta
 
         if ($diff->isNotEmpty()) {
             throw Exception::unknownAttributes($diff->implode('", "'));
-        }
-
-        if (
-            Str::contains($column->get('name'), '.')
-            && ($meta->contains('sortable'))
-        ) {
-            throw Exception::unsupported($column->get('name'));
         }
 
         if ($meta->has('filterable') && $meta->has('icon')) {
